Inicjalny sklep + bind-mount

+++ prestashop-docker-project-1/prestashop/classes/Manufacturer.php
@@ -24,8 +24,6 @@
  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
  */
 
-use PrestaShopBundle\Form\Admin\Type\FormattedTextareaType;
-
 /**
  * Class ManufacturerCore.
  */
@@ -36,10 +34,10 @@ class ManufacturerCore extends ObjectModel
     /** @var string Name */
     public $name;
 
-    /** @var string|array<int, string> Description */
+    /** @var array<string> Description */
     public $description;
 
-    /** @var string|array<int, string> Short description */
+    /** @var array<string> Short description */
     public $short_description;
 
     /** @var int Address */
@@ -54,10 +52,13 @@ class ManufacturerCore extends ObjectModel
     /** @var string Friendly URL */
     public $link_rewrite;
 
-    /** @var string|array<int, string> Meta title */
+    /** @var array<string> Meta title */
     public $meta_title;
 
-    /** @var string|array<int, string> Meta description */
+    /** @var array<string> Meta keywords */
+    public $meta_keywords;
+
+    /** @var array<string> Meta description */
     public $meta_description;
 
     /** @var bool active */
@@ -77,10 +78,11 @@ class ManufacturerCore extends ObjectModel
             'date_upd' => ['type' => self::TYPE_DATE],
 
             /* Lang fields */
-            'description' => ['type' => self::TYPE_HTML, 'lang' => true, 'validate' => 'isCleanHtml', 'size' => FormattedTextareaType::LIMIT_MEDIUMTEXT_UTF8_MB4],
-            'short_description' => ['type' => self::TYPE_HTML, 'lang' => true, 'validate' => 'isCleanHtml', 'size' => FormattedTextareaType::LIMIT_MEDIUMTEXT_UTF8_MB4],
+            'description' => ['type' => self::TYPE_HTML, 'lang' => true, 'validate' => 'isCleanHtml'],
+            'short_description' => ['type' => self::TYPE_HTML, 'lang' => true, 'validate' => 'isCleanHtml'],
             'meta_title' => ['type' => self::TYPE_STRING, 'lang' => true, 'validate' => 'isGenericName', 'size' => 255],
             'meta_description' => ['type' => self::TYPE_STRING, 'lang' => true, 'validate' => 'isGenericName', 'size' => 512],
+            'meta_keywords' => ['type' => self::TYPE_STRING, 'lang' => true, 'validate' => 'isGenericName'],
         ],
     ];
 
@@ -134,8 +136,6 @@ class ManufacturerCore extends ObjectModel
 
             return $this->deleteImage();
         }
-
-        return false;
     }
 
     /**
@@ -143,12 +143,16 @@ class ManufacturerCore extends ObjectModel
      *
      * return boolean Deletion result
      */
-    public function deleteSelection(array $selection)
+    public function deleteSelection($selection)
     {
+        if (!is_array($selection)) {
+            die(Tools::displayError());
+        }
+
         $result = true;
         foreach ($selection as $id) {
             $this->id = (int) $id;
-            $this->id_address = static::getManufacturerAddress();
+            $this->id_address = Manufacturer::getManufacturerAddress();
             $result = $result && $this->delete();
         }
 
@@ -158,7 +162,7 @@ class ManufacturerCore extends ObjectModel
     /**
      * Get Manufacturer Address ID.
      *
-     * @return bool|int
+     * @return bool|false|string|null
      */
     protected function getManufacturerAddress()
     {
@@ -166,9 +170,7 @@ class ManufacturerCore extends ObjectModel
             return false;
         }
 
-        return (int) Db::getInstance(_PS_USE_SQL_SLAVE_)->getValue(
-            'SELECT `id_address` FROM ' . _DB_PREFIX_ . 'address WHERE `id_manufacturer` = ' . (int) $this->id
-        );
+        return Db::getInstance(_PS_USE_SQL_SLAVE_)->getValue('SELECT `id_address` FROM ' . _DB_PREFIX_ . 'address WHERE `id_manufacturer` = ' . (int) $this->id);
     }
 
     /**
@@ -177,11 +179,11 @@ class ManufacturerCore extends ObjectModel
      * @param bool $getNbProducts [optional] return products numbers for each
      * @param int $idLang Language ID
      * @param bool $active
-     * @param int|bool $p
-     * @param int|bool $n
+     * @param int $p
+     * @param int $n
      * @param bool $allGroup
      *
-     * @return array|bool Manufacturers
+     * @return array Manufacturers
      */
     public static function getManufacturers($getNbProducts = false, $idLang = 0, $active = true, $p = false, $n = false, $allGroup = false, $group_by = false, $withProduct = false)
     {
@@ -229,7 +231,7 @@ class ManufacturerCore extends ObjectModel
 						WHERE p.`id_product` = cp.`id_product` AND cg.`id_group` ' . $sqlGroups . '
 					)') . '
 					GROUP BY p.`id_manufacturer`'
-            );
+                );
 
             $counts = [];
             foreach ($results as $result) {
@@ -248,7 +250,7 @@ class ManufacturerCore extends ObjectModel
         $totalManufacturers = count($manufacturers);
         $rewriteSettings = (int) Configuration::get('PS_REWRITING_SETTINGS');
         for ($i = 0; $i < $totalManufacturers; ++$i) {
-            $manufacturers[$i]['link_rewrite'] = ($rewriteSettings ? Tools::str2url($manufacturers[$i]['name']) : 0);
+            $manufacturers[$i]['link_rewrite'] = ($rewriteSettings ? Tools::link_rewrite($manufacturers[$i]['name']) : 0);
         }
 
         return $manufacturers;
@@ -345,7 +347,7 @@ class ManufacturerCore extends ObjectModel
      */
     public function getLink()
     {
-        return Tools::str2url($this->name);
+        return Tools::link_rewrite($this->name);
     }
 
     /**
@@ -355,14 +357,14 @@ class ManufacturerCore extends ObjectModel
      * @param int $idLang
      * @param int $p
      * @param int $n
-     * @param string|null $orderBy
-     * @param string|null $orderWay
+     * @param null $orderBy
+     * @param null $orderWay
      * @param bool $getTotal
      * @param bool $active
      * @param bool $activeCategory
      * @param Context|null $context
      *
-     * @return array|bool|int
+     * @return array|bool
      */
     public static function getProducts(
         $idManufacturer,
@@ -374,7 +376,7 @@ class ManufacturerCore extends ObjectModel
         $getTotal = false,
         $active = true,
         $activeCategory = true,
-        ?Context $context = null
+        Context $context = null
     ) {
         if (!$context) {
             $context = Context::getContext();
@@ -398,7 +400,7 @@ class ManufacturerCore extends ObjectModel
         }
 
         if (!Validate::isOrderBy($orderBy) || !Validate::isOrderWay($orderWay)) {
-            throw new PrestaShopException('Invalid sorting parameters provided.');
+            die(Tools::displayError());
         }
 
         $groups = FrontController::getCurrentCustomerGroups();
@@ -439,15 +441,13 @@ class ManufacturerCore extends ObjectModel
             $alias = 'm.';
         } elseif ($orderBy == 'quantity') {
             $alias = 'stock.';
-        } elseif ($orderBy == 'sales') {
-            $alias = '';
         } else {
             $alias = 'p.';
         }
 
         $sql = 'SELECT p.*, product_shop.*, stock.out_of_stock, IFNULL(stock.quantity, 0) as quantity'
             . (Combination::isFeatureActive() ? ', product_attribute_shop.minimal_quantity AS product_attribute_minimal_quantity, IFNULL(product_attribute_shop.`id_product_attribute`,0) id_product_attribute' : '') . '
-			, pl.`description`, pl.`description_short`, pl.`link_rewrite`, pl.`meta_description`,
+			, pl.`description`, pl.`description_short`, pl.`link_rewrite`, pl.`meta_description`, pl.`meta_keywords`,
 			pl.`meta_title`, pl.`name`, pl.`available_now`, pl.`available_later`, image_shop.`id_image` id_image, il.`legend`, m.`name` AS manufacturer_name,
 				DATEDIFF(
 					product_shop.`date_add`,
@@ -455,7 +455,7 @@ class ManufacturerCore extends ObjectModel
 						"' . date('Y-m-d') . ' 00:00:00",
 						INTERVAL ' . (Validate::isUnsignedInt(Configuration::get('PS_NB_DAYS_NEW_PRODUCT')) ? Configuration::get('PS_NB_DAYS_NEW_PRODUCT') : 20) . ' DAY
 					)
-				) > 0 AS new, psales.`quantity` as sales'
+				) > 0 AS new'
             . ' FROM `' . _DB_PREFIX_ . 'product` p
 			' . Shop::addSqlAssociation('product', 'p') .
             (Combination::isFeatureActive() ? 'LEFT JOIN `' . _DB_PREFIX_ . 'product_attribute_shop` product_attribute_shop
@@ -466,8 +466,6 @@ class ManufacturerCore extends ObjectModel
 					ON (image_shop.`id_product` = p.`id_product` AND image_shop.cover=1 AND image_shop.id_shop=' . (int) $context->shop->id . ')
 			LEFT JOIN `' . _DB_PREFIX_ . 'image_lang` il
 				ON (image_shop.`id_image` = il.`id_image` AND il.`id_lang` = ' . (int) $idLang . ')
-            LEFT JOIN `' . _DB_PREFIX_ . 'product_sale` psales
-					ON psales.`id_product` = p.`id_product`
 			LEFT JOIN `' . _DB_PREFIX_ . 'manufacturer` m
 				ON (m.`id_manufacturer` = p.`id_manufacturer`)
 			' . Product::sqlStock('p', 0);
@@ -505,7 +503,7 @@ class ManufacturerCore extends ObjectModel
             $result = array_slice($result, (int) (($p - 1) * $n), (int) $n);
         }
 
-        return $result;
+        return Product::getProductsProperties($idLang, $result);
     }
 
     /**
